feat(hooks): model reducer and assorted hooks
useReducer is a perfect fit for progressively mutating a model.

Borrowed redux interfaces for creating actions and the reducer. Probably
less robust than directly including them. Redux may be an appropriate
substitute for our pure react reducer, but that requires further
evaluation.

Biggest addition is probably useKey, which takes a key and fetches the
associated schema and model; it also prepares action dispatchers for
mutating said value.

useMappedField and useLocalizedString are just short-hand hooks for a
common pattern