fix(hooks): memoize defaultForSchema to prevent looping
React was crashing the call stack on some array examples; likely would
have affected object examples, too. This is because of the way the new
zustand's useShallow helper determines if a re-render is necessary.
defaultForSchema would return a new array on every call, which would
not be identical to the previous value.

Memoization based on WeakMap actually makes perfect sense here.
defaultForSchema accepts an object and calculates the default value --
if we cache that result in a WeakMap, the result will stick around for
exactly as long as the schema.

Of course, this still leaves open the possibility of the caller
returning a new schema object, which could still trigger this behavior.