feat(core): improve array performance using ref for keys
Using a ref combined with a memo -- the ref allows persistent storage,
while the memo ensures that we add or remove keys when the underlying
value is changed. Tried useEffect for that initially, but since it
triggers post-render, the change to the key array isn't noted until the
*next* render. useMemo does not have that problem, as it runs during the
render, and we can still get our desired side effect out of it.