feat(hooks): optimizations for array elements and manipulation
Removes the need for passing the volatile index of a given form element
around, thus reducing the rendering complexity of array element deletion
and reordering.

Rather than relying on the index, under the new model, array elements
are given an "id" value which can be used both as a React key (providing
stable rendering) and as a stand-in for the index value when performing
mutations (like removing an element, reordering elements, or updating an
element's value). While this does mean we must do some extra work look
up the index value in every place where we need it, distributing this
work to those points means that our React components are relatively
insulated from the volatility of those indexes. Removing an element at
the start of the array will no longer force every subsequent array
element to re-render.